Revert "Bad"
This reverts commit 02e14e703bf2309544d2c41cead9c24ce5ab401f.

+++ Assets/Scripts/PlayerScript.cs
@@ -11,8 +11,8 @@ public class PlayerScript : MonoBehaviour
     public Transform firePoint;
 
     [Header("Player Parameters")]
-    public float runSpeed = 10;
-    public float crouchSpeed = 5;
+    public int runSpeed = 10;
+    public int crouchSpeed = 5;
     public Material lineMaterial;
     public GameObject circlePrefab;
